* upstream/master: (637 commits)
  Update changelog
  update(changelog)
  fix(linux-build): Update rpm spec to use new package name
  fix(cli): s/n1/nylas-mail/
  fix(linux-build): Change package name to Nylas Mail
  fix(SFDC): fix React error in tokenizing text field
  fix(watchman): There's some stuff in build/ we don't want to ignore
  fix(spec): re-enable NylasEnv specs and fix duplicate rejection log
  :art:(backoff-scheduler) Add getters
  bump(version): 1.0.28
  update(changelog)
  bump(k2)
  [battery] Add BatteryStatusManager
  bump(K2)
  fix(exports) Add backoff schedulers
  bump(k2)
  feat(offline) Re add offline status notification
  bump(k2): SFDC
  reafctor(scheduler): Move SearchIndexer -> SearchIndexScheduler
  feat(backoff-scheduler): Add a backoff scheduler service
  ...

+++ src/flux/stores/file-download-store.es6
@@ -9,6 +9,7 @@ import progress from 'request-progress';
 import NylasStore from 'nylas-store';
 import Actions from '../actions';
 import NylasAPI from '../nylas-api';
+import NylasAPIRequest from '../nylas-api-request';
 
 
 Promise.promisifyAll(fs);
@@ -44,13 +45,18 @@ const THUMBNAIL_WIDTH = 320
 export class Download {
   static State = State
 
-  constructor({accountId, fileId, targetPath, filename, filesize, progressCallback}) {
+  constructor({accountId, fileId, targetPath, filename, filesize, progressCallback, retryWithBackoff}) {
     this.accountId = accountId;
     this.fileId = fileId;
     this.targetPath = targetPath;
     this.filename = filename;
     this.filesize = filesize;
     this.progressCallback = progressCallback;
+    this.retryWithBackoff = retryWithBackoff || false;
+    this.timeout = 15000;
+    this.maxTimeout = 2 * 60 * 1000;
+    this.attempts = 0;
+    this.maxAttempts = 10;
     if (!this.accountId) {
       throw new Error("Download.constructor: You must provide a non-empty accountId.");
     }
@@ -93,14 +99,22 @@ export class Download {
       const stream = fs.createWriteStream(this.targetPath);
       this.state = State.Downloading;
 
+      let startRequest = null;
+
       const onFailed = (err) => {
         this.request = null;
         stream.end();
-        this.state = State.Failed;
-        if (fs.existsSync(this.targetPath)) {
-          fs.unlinkSync(this.targetPath);
+        if (!this.retryWithBackoff || this.attempts >= this.maxAttempts) {
+          this.state = State.Failed;
+          if (fs.existsSync(this.targetPath)) {
+            fs.unlinkSync(this.targetPath);
+          }
+          reject(err);
+          return;
         }
-        reject(err);
+
+        this.timeout = Math.min(this.maxTimeout, this.timeout * 2);
+        startRequest();
       };
 
       const onSuccess = () => {
@@ -111,39 +125,51 @@ export class Download {
         resolve(this);
       };
 
-      NylasAPI.makeRequest({
-        json: false,
-        path: `/files/${this.fileId}/download`,
-        accountId: this.accountId,
-        encoding: null, // Tell `request` not to parse the response data
-        started: (req) => {
-          this.request = req;
-          return progress(this.request, {throtte: 250})
-          .on('progress', (prog) => {
-            this.percent = prog.percent;
-            this.progressCallback();
-          })
+      startRequest = () => {
+        console.info(`starting download with ${this.timeout}ms timeout`);
+        const request = new NylasAPIRequest({
+          api: NylasAPI,
+          options: {
+            json: false,
+            path: `/files/${this.fileId}/download`,
+            accountId: this.accountId,
+            encoding: null, // Tell `request` not to parse the response data
+            timeout: this.timeout,
+            started: (req) => {
+              this.attempts += 1;
+              this.request = req;
+              return progress(this.request, {throtte: 250})
+              .on('progress', (prog) => {
+                this.percent = prog.percent;
+                this.progressCallback();
+              })
 
-          // This is a /socket/ error event, not an HTTP error event. It fires
-          // when the conn is dropped, user if offline, but not on HTTP status codes.
-          // It is sometimes called in place of "end", not before or after.
-          .on('error', onFailed)
+              // This is a /socket/ error event, not an HTTP error event. It fires
+              // when the conn is dropped, user if offline, but not on HTTP status codes.
+              // It is sometimes called in place of "end", not before or after.
+              .on('error', onFailed)
 
-          .on('end', () => {
-            if (this.state === State.Failed) { return; }
+              .on('end', () => {
+                if (this.state === State.Failed) { return; }
 
-            const {response} = this.request
-            const statusCode = response ? response.statusCode : null;
-            if ([200, 202, 204].includes(statusCode)) {
-              onSuccess();
-            } else {
-              onFailed(new Error(`Server returned a ${statusCode}`));
-            }
-          })
+                const {response} = this.request
+                const statusCode = response ? response.statusCode : null;
+                if ([200, 202, 204].includes(statusCode)) {
+                  onSuccess();
+                } else {
+                  onFailed(new Error(`Server returned a ${statusCode}`));
+                }
+              })
 
-          .pipe(stream);
-        },
-      });
+              .pipe(stream);
+            },
+          },
+        });
+
+        request.run()
+      };
+
+      startRequest();
     });
     return this.promise
   }
@@ -165,7 +191,6 @@ class FileDownloadStore extends NylasStore {
     this.listenTo(Actions.fetchAndSaveFile, this._fetchAndSave);
     this.listenTo(Actions.fetchAndSaveAllFiles, this._fetchAndSaveAll);
     this.listenTo(Actions.abortFetchFile, this._abortFetchFile);
-    this.listenTo(Actions.didPassivelyReceiveNewModels, this._onNewMailReceived);
 
     this._downloads = {};
     this._filePreviewPaths = {};
@@ -210,16 +235,6 @@ class FileDownloadStore extends NylasStore {
     return this._filePreviewPaths[fileId];
   }
 
-  _onNewMailReceived = (incoming) => {
-    if (NylasEnv.config.get('core.attachments.downloadPolicy') !== 'on-receive') {
-      return;
-    }
-    if (!incoming.message) { return; }
-    incoming.message.forEach((message) => {
-      message.files.forEach((file) => this._fetch(file));
-    })
-  }
-
   // Returns a promise with a Download object, allowing other actions to be
   // daisy-chained to the end of the download operation.
   _runDownload(file) {
@@ -238,6 +253,7 @@ class FileDownloadStore extends NylasStore {
       filename: file.displayName(),
       targetPath,
       progressCallback: () => this.trigger(),
+      retryWithBackoff: true,
     });
 
     // Do we actually need to queue and run the download? Queuing a download
